+++ app/Providers/AppServiceProvider.php
@@ -40,11 +40,10 @@ class AppServiceProvider extends ServiceProvider
         FacadesDB::listen(function ($query) {
             if (explode(" ", $query->sql)[0] !== 'select')
             {
-                //dd($query->bindings);
                 File::append(
                     storage_path('/logs/query.log'),
                     '[' . date('Y-m-d H:i:s') . ']' . PHP_EOL .
-                    '[' . auth()->user()->id . ' , ' . auth()->user()->name . ']' . PHP_EOL . $query->sql . ' [' . $this->implotar(', ', $query->bindings) . ']' . PHP_EOL . PHP_EOL
+                    '[' . (auth()->user()->id ?? 'null') . ' , ' . (auth()->user()->name ?? 'Sin Usuario Logueado') . ']' . PHP_EOL . $query->sql . ' [' . $this->implotar(', ', $query->bindings) . ']' . PHP_EOL . PHP_EOL
                 );
             }
         });