+++ database/migrations/2021_02_19_003043_add_us_pass_equipos_table.php
@@ -1,0 +1,34 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class AddUsPassEquiposTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('equipos', function (Blueprint $table) {
+            $table->string('usuario', 255)->after('comentario')->nullable();
+            $table->string('password', 255)->after('comentario')->nullable();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('equipos', function (Blueprint $table) {
+            $table->dropColumn('usuario');
+            $table->dropColumn('password');
+        });
+    }
+}