+++ resources/views/agregarPanel.blade.php
@@ -46,7 +46,7 @@ $mostrarSololectura = true;
                 <div class="form-group col-md-3">
                     <label for="panel_ant">Panel Anterior: </label>
                     <select class="form-control" name="panel_ant">
-                        <option value="">Gateway...</option>
+                        <option value="">Internet...</option>
                         @foreach ($paneles as $panel)
                             @if ($panel->activo)
                                 <option value="{{$panel->id}}">{{$panel->relEquipo->nombre}}->{{$panel->relEquipo->ip}}</option>';