+++ resources/views/layouts/footer.blade.php
@@ -1,5 +1,5 @@
     <footer class="bg-dark fixed-bottom text-light text-center p-2">
-        Developed by Palky Inc. &copy;2020. Versión {{Config::get('constants.SLAM_VERSION')}}
+        &copy;Bootstrap.4.5.3 &copy;Laravel.v8 &copy;Vue.JS.2.6.12 | Developed by Palky Inc. &copy;Slam.v{{Config::get('constants.SLAM_VERSION')}} - 2020
     </footer>
 </body>
 </html>